+++ Source/JavaScriptCore/bytecode/InlineCacheCompiler.h
@@ -20,7 +20,7 @@
  * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
  * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. 
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
 #pragma once
@@ -43,6 +43,7 @@ class GetterSetter;
 class CCallHelpers;
 class CodeBlock;
 class PolymorphicAccess;
+class ProxyObjectAccessCase;
 class StructureStubInfo;
 class WatchpointsOnStructureStubInfo;
 
@@ -58,19 +59,19 @@ public:
         GeneratedFinalCode, // Generated so much code that we never want to generate code again.
         ResetStubAndFireWatchpoints // We found out some data that makes us want to start over fresh with this stub. Currently, this happens when we detect poly proto.
     };
-    
+
 
     AccessGenerationResult() = default;
     AccessGenerationResult(AccessGenerationResult&&) = default;
     AccessGenerationResult& operator=(AccessGenerationResult&&) = default;
-    
+
     AccessGenerationResult(Kind kind)
         : m_kind(kind)
     {
         RELEASE_ASSERT(kind != GeneratedNewCode);
         RELEASE_ASSERT(kind != GeneratedFinalCode);
     }
-    
+
     AccessGenerationResult(Kind kind, CodePtr<JITStubRoutinePtrTag> code)
         : m_kind(kind)
         , m_code(code)
@@ -78,39 +79,39 @@ public:
         RELEASE_ASSERT(kind == GeneratedNewCode || kind == GeneratedFinalCode);
         RELEASE_ASSERT(code);
     }
-    
+
     bool operator==(const AccessGenerationResult& other) const
     {
         return m_kind == other.m_kind && m_code == other.m_code;
     }
-    
+
     bool operator!=(const AccessGenerationResult& other) const
     {
         return !(*this == other);
     }
-    
+
     explicit operator bool() const
     {
         return *this != AccessGenerationResult();
     }
-    
+
     Kind kind() const { return m_kind; }
-    
+
     const CodePtr<JITStubRoutinePtrTag>& code() const { return m_code; }
-    
+
     bool madeNoChanges() const { return m_kind == MadeNoChanges; }
     bool gaveUp() const { return m_kind == GaveUp; }
     bool buffered() const { return m_kind == Buffered; }
     bool generatedNewCode() const { return m_kind == GeneratedNewCode; }
     bool generatedFinalCode() const { return m_kind == GeneratedFinalCode; }
     bool shouldResetStubAndFireWatchpoints() const { return m_kind == ResetStubAndFireWatchpoints; }
-    
+
     // If we gave up on this attempt to generate code, or if we generated the "final" code, then we
     // should give up after this.
     bool shouldGiveUpNow() const { return gaveUp() || generatedFinalCode(); }
-    
+
     bool generatedSomeCode() const { return generatedNewCode() || generatedFinalCode(); }
-    
+
     void dump(PrintStream&) const;
 
     void addWatchpointToFire(InlineWatchpointSet& set, StringFireDetail detail)
@@ -123,7 +124,7 @@ public:
         for (auto& pair : m_watchpointsToFire)
             pair.first.invalidate(vm, pair.second);
     }
-    
+
 private:
     Kind m_kind;
     CodePtr<JITStubRoutinePtrTag> m_code;
@@ -134,6 +135,8 @@ class PolymorphicAccess {
     WTF_MAKE_NONCOPYABLE(PolymorphicAccess);
     WTF_MAKE_STRUCT_FAST_ALLOCATED_WITH_HEAP_IDENTIFIER(PolymorphicAccess);
 public:
+    friend class InlineCacheCompiler;
+
     PolymorphicAccess();
     ~PolymorphicAccess();
 
@@ -144,9 +147,7 @@ public:
 
     AccessGenerationResult addCase(
         const GCSafeConcurrentJSLocker&, VM&, CodeBlock*, StructureStubInfo&, Ref<AccessCase>);
-    
-    AccessGenerationResult regenerate(const GCSafeConcurrentJSLocker&, VM&, JSGlobalObject*, CodeBlock*, ECMAMode, StructureStubInfo&);
-    
+
     bool isEmpty() const { return m_list.isEmpty(); }
     unsigned size() const { return m_list.size(); }
     const AccessCase& at(unsigned i) const { return *m_list[i]; }
@@ -156,7 +157,7 @@ public:
 
     // If this returns false then we are requesting a reset of the owning StructureStubInfo.
     bool visitWeak(VM&) const;
-    
+
     // This returns true if it has marked everything it will ever marked. This can be used as an
     // optimization to then avoid calling this method again during the fixpoint.
     template<typename Visitor> void propagateTransitions(Visitor&) const;
@@ -165,7 +166,7 @@ public:
 
     void dump(PrintStream& out) const;
     bool containsPC(void* pc) const
-    { 
+    {
         if (!m_stubRoutine)
             return false;
 
@@ -176,44 +177,24 @@ public:
 private:
     friend class AccessCase;
     friend class CodeBlock;
-    friend struct AccessGenerationState;
-    
+    friend class InlineCacheCompiler;
+
     typedef Vector<RefPtr<AccessCase>, 2> ListType;
-    
-    void commit(
-        const GCSafeConcurrentJSLocker&, VM&, std::unique_ptr<WatchpointsOnStructureStubInfo>&, CodeBlock*, StructureStubInfo&,
-        AccessCase&);
 
     ListType m_list;
     RefPtr<PolymorphicAccessJITStubRoutine> m_stubRoutine;
     std::unique_ptr<WatchpointsOnStructureStubInfo> m_watchpoints;
 };
 
-struct AccessGenerationState {
-    AccessGenerationState(VM& vm, JSGlobalObject* globalObject, ECMAMode ecmaMode)
-        : m_vm(vm) 
+class InlineCacheCompiler {
+public:
+    InlineCacheCompiler(VM& vm, JSGlobalObject* globalObject, ECMAMode ecmaMode, StructureStubInfo& stubInfo)
+        : m_vm(vm)
         , m_globalObject(globalObject)
+        , m_stubInfo(&stubInfo)
         , m_ecmaMode(ecmaMode)
     {
     }
-    VM& m_vm;
-    JSGlobalObject* const m_globalObject;
-    CCallHelpers* jit { nullptr };
-    ScratchRegisterAllocator* allocator;
-    ScratchRegisterAllocator::PreservedState preservedReusedRegisterState;
-    PolymorphicAccess* access { nullptr };
-    StructureStubInfo* stubInfo { nullptr };
-    MacroAssembler::JumpList success;
-    MacroAssembler::JumpList failAndRepatch;
-    MacroAssembler::JumpList failAndIgnore;
-    GPRReg scratchGPR { InvalidGPRReg };
-    FPRReg scratchFPR { InvalidFPRReg };
-    const ECMAMode m_ecmaMode { ECMAMode::sloppy() };
-    std::unique_ptr<WatchpointsOnStructureStubInfo> watchpoints;
-    Vector<StructureID> weakStructures;
-    Bag<OptimizingCallLinkInfo> m_callLinkInfos;
-    bool m_doesJSCalls : 1 { false };
-    bool m_doesCalls : 1 { false };
 
     void installWatchpoint(CodeBlock*, const ObjectPropertyCondition&);
 
@@ -251,7 +232,7 @@ struct AccessGenerationState {
 
     bool needsToRestoreRegistersIfException() const { return m_needsToRestoreRegistersIfException; }
     CallSiteIndex originalCallSiteIndex() const;
-    
+
     void emitExplicitExceptionHandler();
 
     void setSpillStateForJSCall(SpillState& spillState)
@@ -266,9 +247,45 @@ struct AccessGenerationState {
 
     ScratchRegisterAllocator makeDefaultScratchAllocator(GPRReg extraToLock = InvalidGPRReg);
 
+    // Fall through on success. Two kinds of failures are supported: fall-through, which means that we
+    // should try a different case; and failure, which means that this was the right case but it needs
+    // help from the slow path.
+    void generateWithGuard(AccessCase&, MacroAssembler::JumpList& fallThrough);
+
+    // Fall through on success, add a jump to the failure list on failure.
+    void generate(AccessCase&);
+
+    void generateImpl(AccessCase&);
+
+    static bool canEmitIntrinsicGetter(StructureStubInfo&, JSFunction*, Structure*);
+
+    VM& vm() { return m_vm; }
+
+    AccessGenerationResult regenerate(const GCSafeConcurrentJSLocker&, PolymorphicAccess&, CodeBlock*);
+
 private:
     const ScalarRegisterSet& liveRegistersToPreserveAtExceptionHandlingCallSite();
-    
+
+    void emitDOMJITGetter(GetterSetterAccessCase&, const DOMJIT::GetterSetter*, GPRReg baseForGetGPR);
+    void emitModuleNamespaceLoad(ModuleNamespaceAccessCase&, MacroAssembler::JumpList& fallThrough);
+    void emitProxyObjectAccess(ProxyObjectAccessCase&, MacroAssembler::JumpList& fallThrough);
+    void emitIntrinsicGetter(IntrinsicGetterAccessCase&);
+
+    VM& m_vm;
+    JSGlobalObject* const m_globalObject;
+    StructureStubInfo* m_stubInfo { nullptr };
+    const ECMAMode m_ecmaMode { ECMAMode::sloppy() };
+    CCallHelpers* m_jit { nullptr };
+    ScratchRegisterAllocator* m_allocator { nullptr };
+    MacroAssembler::JumpList m_success;
+    MacroAssembler::JumpList m_failAndRepatch;
+    MacroAssembler::JumpList m_failAndIgnore;
+    ScratchRegisterAllocator::PreservedState m_preservedReusedRegisterState;
+    std::unique_ptr<WatchpointsOnStructureStubInfo> m_watchpoints;
+    GPRReg m_scratchGPR { InvalidGPRReg };
+    FPRReg m_scratchFPR { InvalidFPRReg };
+    Vector<StructureID> m_weakStructures;
+    Bag<OptimizingCallLinkInfo> m_callLinkInfos;
     ScalarRegisterSet m_liveRegistersToPreserveAtExceptionHandlingCallSite;
     ScalarRegisterSet m_liveRegistersForCall;
     CallSiteIndex m_callSiteIndex;
@@ -276,6 +293,8 @@ private:
     bool m_calculatedRegistersForCallAndExceptionHandling : 1 { false };
     bool m_needsToRestoreRegistersIfException : 1 { false };
     bool m_calculatedCallSiteIndex : 1 { false };
+    bool m_doesJSCalls : 1 { false };
+    bool m_doesCalls : 1 { false };
 };
 
 } // namespace JSC
