+++ Source/JavaScriptCore/bytecode/StructureStubInfo.h
@@ -28,11 +28,11 @@
 #include "CacheableIdentifier.h"
 #include "CodeBlock.h"
 #include "CodeOrigin.h"
+#include "InlineCacheCompiler.h"
 #include "Instruction.h"
 #include "JITStubRoutine.h"
 #include "MacroAssembler.h"
 #include "Options.h"
-#include "PolymorphicAccess.h"
 #include "PutKind.h"
 #include "RegisterSet.h"
 #include "Structure.h"
